Added setup.py
Added a setup.py file to support building. Required for this was to make
a python package and update all of the imports in the sources. Now, it
must be run as `python -m cumulative_snow.main`, but the setup.py dist
takes care of this for you.

Also added xar as a dev dep so you can build a xar file to run with:
```
python setup.py bdist_xar
./dist/cumulative_snow.xar --help
```